% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_all_match_sets.R
\name{resolve_all_match_sets}
\alias{resolve_all_match_sets}
\title{Resolve match (list of source-match sets).}
\usage{
resolve_all_match_sets(match_set_list)
}
\arguments{
\item{match_set_list}{}
}
\value{
Returns a list with two elements. "new_set_list" is a cleaned version
of match_set_list. It contains the same number of elements but with the
sources and match terms selected by resolve_match_set. "auto_accept_index" is
a Boolean index indicating which of the elements in new_set_index can be
assumed to be correct matches (and thus do not need manual user review).
}
\description{
Attempts to resolve each of a collection of match sets.
}
\details{
Takes a list of match sets, applies resolve_match_set to each, and gathers
the results up in a list whose features can be fed to the
process_fuzzy_matches function.
}

