% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_match.R
\name{resolve_match}
\alias{resolve_match}
\title{Resolve match (source and one match).}
\usage{
resolve_match(source_term, match_term)
}
\arguments{
\item{source_term}{The term selected for testing against similar terms.}

\item{match_term}{A term identified as similar.}
}
\value{
Returns a list specifying which term was selected as the source
("source"), which term (if either) was selected as the match ("match"), and
whether the pair was considered auto-resolved or needs manual attention
("auto_accept").
}
\description{
Inspects source-match pair to see if can auto-determine if valid match or
not.
}
\details{
The heart of the auto-resolver, this function contains the logic for
evaluating a single source--match pair of strings.
}

