% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_match_set.R
\name{resolve_match_set}
\alias{resolve_match_set}
\title{Resolve match (source and all matches).}
\usage{
resolve_match_set(match_set)
}
\arguments{
\item{match_set}{}
}
\value{
Returns a list with the same components as resolve_match ("source",
"match", "auto_accept"). However, "match" now contains all retained matches
and "auto_accept" is an evaluation of the set rather than a single pair.
}
\description{
Evaluates a suggested source term and its collection of matches.
}
\details{
Takes a match-set - as defined by the fuzzy match function - and evaluates
all the pairings. Attempts to choose the best source and matches and then
evaluates whether it is appropriate to consider the set auto-resolved.
}

