% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_elements.R
\name{sort_elements}
\alias{sort_elements}
\title{Sort complex elements.}
\usage{
sort_elements(tar_vector, split_reg = "----", decreasing = FALSE,
  dedup = TRUE, collapse_first = TRUE, cf_split = split_reg)
}
\arguments{
\item{tar_vector}{The vector of strings with elements to sort.}

\item{split_reg}{String providing the regex logic for breaking/collapsing
target strings.}

\item{decreasing}{Boolean controlling sort direction.}
}
\value{
Returns an updated vector.
}
\description{
Sorts (and optionally deduplicates) the elements of complex field.
}

